proxied/proxy/requirements.txt
Flask==0.12.2
PyJWT==1.5.3
cryptography==2.1.4
requests==2.18.4
